Updated RegEx usage and validator name
